refactor: changed import path to path alias
