+++ src/App.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom'
 
-import './styles/app.scss'
+import '@/styles/app.scss'
 
-import { NavBar, Modal } from './components/shared'
-import { Home, Algorithms, About } from './pages'
+import { NavBar, Modal } from '@/components/shared'
+import { Home, Algorithms, About } from '@/pages'
 
 const App: React.FC = () => {
   return (
