+++ src/components/Home/SidePanel.tsx
@@ -1,8 +1,9 @@
 import React, { ReactEventHandler, useState } from 'react'
 import { useDispatch } from 'react-redux'
 
-import { changeAlgorithm } from '../../store/slice/sorting'
-import { Header, Dropdown } from '../shared'
+import { changeAlgorithm } from '@/store/slice/sorting'
+
+import { Header, Dropdown } from '@/components/shared'
 
 // WIP
