+++ src/components/shared/Modal.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useEffect, useState } from 'react';
+import { FC, useEffect, useState } from 'react';
 
 const localStorageKey = 'modal-has-been-shown'
